requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-cors-headers
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
oauthlib
pycparser
PyJWT
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
tzdata
urllib3
